:truck: RESTRUCTURE: Move assets into separate folders and rename assets import paths
