+++ assets/icons/star.js
