+++ assets/images/icon@2x.png
